+++ src/make_cmake_files.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-
 #
 #  PCMSolver, an API for the Polarizable Continuum Model
 #  Copyright (C) 2018 Roberto Di Remigio, Luca Frediani and contributors.
@@ -23,9 +21,6 @@
 #  PCMSolver API, see: <http://pcmsolver.readthedocs.io/>
 #
 
-# -*- python -*-
-# -*- coding: utf-8 -*-
-# vim:filetype=python:
 # Create CMakeLists.txt template for leaf directories
 # (c) Roberto Di Remigio  <roberto.d.remigio@uit.no>
 # licensed under the GNU Lesser General Public License
@@ -33,8 +28,6 @@
 import os
 import sys
 import glob
-
-sys.path.append('cmake/lib/docopt')
 import docopt
 
 options = """
@@ -127,8 +120,15 @@ properties = """set_target_properties({0}
     CXX_VISIBILITY_PRESET hidden
     VISIBILITY_INLINES_HIDDEN 1
   )
+target_compile_definitions({0}
+  PUBLIC
+    PCMSolver_EXPORTS
+  )
 """.format(libname)
-cxx_compile_options = """add_dependencies({0} generate-config-hpp)
+cxx_compile_options = """if(BUILD_CUSTOM_BOOST)
+  add_dependencies({0} custom_boost)
+endif()
+add_dependencies({0} generate-config-hpp)
 target_compile_options({0}
   PRIVATE
     {1}