+++ tests/bi_operators/bi_operators_purisima.cpp
@@ -1,4 +1,4 @@
-/**
+/*
  * PCMSolver, an API for the Polarizable Continuum Model
  * Copyright (C) 2018 Roberto Di Remigio, Luca Frediani and contributors.
  *
@@ -41,8 +41,8 @@
 using namespace pcm;
 using bi_operators::Purisima;
 using cavity::GePolCavity;
-using green::Vacuum;
 using green::UniformDielectric;
+using green::Vacuum;
 
 SCENARIO(
     "A collocation integrator with diagonal elements according to Purisima for D",